Feature: add long data type instead of int

+++ src/utils/philosophers_utils.c
@@ -14,7 +14,7 @@ int is_even(int n)
     return (0);
 }
 
-int set_time()
+long    set_time()
 {
     struct timeval  time_struct;
